Convert docs from Sphinx to Hugo
The Sphinx build was broken for some time now,
Hugo has the advantage of only needing a single
binary, which makes building the docs much easier.

In order to build publish the docs, a GitHub Action
was added. This action pushes the docs to the `gh-pages`
branch so it can be served by GitHub.

Resolves #1225